step3 수정사항 - application 에 있던 계산식 제거 - 메서드 위치 수정 - 클래스 패키지 수정
