+++ src/test/java/lotto/domain/LottoProfitTest.java
@@ -1,4 +1,4 @@
-package lotto.dto;
+package lotto.domain;
 
 import lotto.vo.LottoMatchCount;
 import org.junit.Test;
@@ -10,7 +10,7 @@ public class LottoProfitTest {
 
     @Test
     public void 수익금계산테스트() {
-        LottoProfit lottoProfit = new LottoProfit(3, 2, 5500);
+        LottoProfit lottoProfit = new LottoProfit(2, 5500);
         LottoMatchCount lottoMatchCount = new LottoMatchCount(0, 0, 0, 0, 1);
         lottoProfit.calculateProfit(lottoMatchCount);
         assertThat(lottoProfit.getProfit()).isEqualTo(1.0);
